+++ internal/server/openai_chat.go
@@ -1,21 +1,12 @@
 package server
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"net/http"
-	"strings"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/openai/openai-go/v3"
-	"github.com/openai/openai-go/v3/packages/param"
-	"github.com/openai/openai-go/v3/packages/ssestream"
-	"github.com/openai/openai-go/v3/responses"
-	"github.com/sirupsen/logrus"
 	"github.com/tingly-dev/tingly-box/internal/constant"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/protocol/stream"
@@ -257,393 +248,4 @@ func (s *Server) streamOpenAIChat(c *gin.Context, provider *typ.Provider, origin
 
 	// Track usage from stream handler
 	s.trackUsageWithTokenUsage(c, usage, err)
-}
-
-func (s *Server) handleOpenAIStreamResponse(c *gin.Context, streamResp *ssestream.Stream[openai.ChatCompletionChunk], req *openai.ChatCompletionNewParams, responseModel string, disableStreamUsage bool) {
-	// Accumulate usage from stream chunks
-	var inputTokens, outputTokens int
-	var hasUsage bool
-	var contentBuilder strings.Builder
-	var firstChunkID string // Store the first chunk ID for usage estimation
-
-	defer func() {
-		if r := recover(); r != nil {
-			logrus.Errorf("Panic in streaming handler: %v", r)
-			// Track panic as error with any usage we accumulated
-			if hasUsage {
-				usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, 0)
-				s.trackUsageWithTokenUsage(c, usage, fmt.Errorf("panic: %v", r))
-			}
-			// Try to send an error event if possible
-			if c.Writer != nil {
-				c.Writer.WriteHeader(http.StatusInternalServerError)
-				c.SSEvent("", map[string]interface{}{
-					"error": map[string]interface{}{
-						"message": "Internal streaming error",
-						"type":    "internal_error",
-					},
-				})
-				if flusher, ok := c.Writer.(http.Flusher); ok {
-					flusher.Flush()
-				}
-			}
-		}
-		// Ensure stream is always closed
-		if streamResp != nil {
-			if err := streamResp.Close(); err != nil {
-				logrus.Errorf("Error closing stream: %v", err)
-			}
-		}
-	}()
-
-	// Set SSE headers (mimicking OpenAI response headers)
-	c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Authorization, Cache-Control")
-	c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-	c.Header("Access-Control-Allow-Origin", "*")
-	c.Header("Content-Type", "text/event-stream; charset=utf-8")
-	c.Header("Cache-Control", "no-cache")
-	c.Header("Connection", "keep-alive")
-
-	// Create a flusher to ensure immediate sending of data
-	flusher, ok := c.Writer.(http.Flusher)
-	if !ok {
-		c.JSON(http.StatusInternalServerError, ErrorResponse{
-			Error: ErrorDetail{
-				Message: "Streaming not supported by this connection",
-				Type:    "api_error",
-				Code:    "streaming_unsupported",
-			},
-		})
-		return
-	}
-
-	// Process the stream with context cancellation checking
-	protocol.RunLoop(c, func(w io.Writer) bool {
-		// Check context cancellation first
-		select {
-		case <-c.Request.Context().Done():
-			logrus.Debug("Client disconnected, stopping OpenAI stream")
-			return false
-		default:
-		}
-
-		// Try to get next chunk
-		if !streamResp.Next() {
-			// Stream ended
-			return false
-		}
-
-		chatChunk := streamResp.Current()
-		obfuscationValue := stream.GenerateObfuscationString() // Generate obfuscation value once per stream
-
-		// Store the first chunk ID for usage estimation
-		if firstChunkID == "" && chatChunk.ID != "" {
-			firstChunkID = chatChunk.ID
-		}
-
-		// Accumulate usage from chunks (if present)
-		if chatChunk.Usage.PromptTokens != 0 {
-			inputTokens = int(chatChunk.Usage.PromptTokens)
-			hasUsage = true
-		}
-
-		if chatChunk.Usage.CompletionTokens != 0 {
-			outputTokens = int(chatChunk.Usage.CompletionTokens)
-			hasUsage = true
-		}
-
-		// Check if we have choices and they're not empty
-		if len(chatChunk.Choices) == 0 {
-			return true
-		}
-
-		choice := chatChunk.Choices[0]
-
-		// Accumulate content for estimation
-		if choice.Delta.Content != "" {
-			contentBuilder.WriteString(choice.Delta.Content)
-		}
-
-		// Build delta map - only include non-empty fields to avoid validation errors
-		delta := map[string]interface{}{}
-		if choice.Delta.Role != "" {
-			delta["role"] = choice.Delta.Role
-		}
-		if choice.Delta.Content != "" {
-			delta["content"] = choice.Delta.Content
-		} else {
-			delta["content"] = ""
-		}
-		if choice.Delta.Refusal != "" {
-			delta["refusal"] = choice.Delta.Refusal
-		} else {
-			delta["refusal"] = nil
-		}
-		if choice.Delta.JSON.FunctionCall.Valid() {
-			delta["function_call"] = choice.Delta.FunctionCall
-		}
-		if len(choice.Delta.ToolCalls) > 0 {
-			delta["tool_calls"] = choice.Delta.ToolCalls
-		}
-
-		// Mark TTFT on the first chunk carrying content (text / tool call /
-		// refusal), not the leading role-only delta. MarkFirstToken is idempotent.
-		if choice.Delta.Content != "" || choice.Delta.Refusal != "" || len(choice.Delta.ToolCalls) > 0 || choice.Delta.JSON.FunctionCall.Valid() {
-			protocol.MarkFirstToken(c)
-		}
-
-		finishReason := &choice.FinishReason
-		if finishReason != nil && *finishReason == "" {
-			finishReason = nil
-		}
-
-		// Prepare the chunk in OpenAI format
-		chunk := map[string]interface{}{
-			"id":      chatChunk.ID,
-			"object":  "chat.completion.chunk",
-			"created": chatChunk.Created,
-			"model":   responseModel,
-			"choices": []map[string]interface{}{
-				{
-					"index":         choice.Index,
-					"delta":         delta,
-					"finish_reason": finishReason,
-					"logprobs":      choice.Logprobs,
-				},
-			},
-		}
-
-		// Add usage if present (usually only in the last chunk) and not disabled
-		if !disableStreamUsage && (chatChunk.Usage.PromptTokens != 0 || chatChunk.Usage.CompletionTokens != 0) {
-			chunk["usage"] = chatChunk.Usage
-		}
-
-		// Add system fingerprint if present
-		if chatChunk.SystemFingerprint != "" {
-			chunk["system_fingerprint"] = chatChunk.SystemFingerprint
-		}
-
-		// Add service_tier if present
-		if chatChunk.ServiceTier != "" {
-			chunk["service_tier"] = chatChunk.ServiceTier
-		} else {
-			chunk["service_tier"] = "default"
-		}
-
-		// Add obfuscation if present in extra fields, otherwise use generated value
-		if obfuscationField, ok := chatChunk.JSON.ExtraFields["obfuscation"]; ok && obfuscationField.Valid() {
-			var upstreamObfuscation string
-			if err := json.Unmarshal([]byte(obfuscationField.Raw()), &upstreamObfuscation); err == nil {
-				chunk["obfuscation"] = upstreamObfuscation
-			} else {
-				chunk["obfuscation"] = obfuscationValue
-			}
-		} else {
-			chunk["obfuscation"] = obfuscationValue
-		}
-
-		// Convert to JSON and send as SSE
-		chunkJSON, err := json.Marshal(chunk)
-		if err != nil {
-			logrus.Errorf("Failed to marshal chunk: %v", err)
-			return true // Continue on marshal error
-		}
-
-		// Send the chunk
-		// MENTION: Must keep extra space
-		c.Writer.WriteString(fmt.Sprintf("data: %s\n\n", chunkJSON))
-		flusher.Flush()
-		return true
-	})
-
-	// Check for stream errors
-	if err := streamResp.Err(); err != nil {
-		// Check if it was a client cancellation
-		if errors.Is(err, context.Canceled) {
-			logrus.Debug("OpenAI stream canceled by client")
-			// Estimate usage if we don't have it
-			if !hasUsage {
-				inputTokens, _ = token.EstimateInputTokens(req)
-				outputTokens = token.EstimateOutputTokens(contentBuilder.String())
-			}
-			usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, 0)
-			s.trackUsageWithTokenUsage(c, usage, err)
-			return
-		}
-
-		logrus.Errorf("Stream error: %v", err)
-
-		// If no usage from stream, estimate it
-		if !hasUsage {
-			inputTokens, _ = token.EstimateInputTokens(req)
-			outputTokens = token.EstimateOutputTokens(contentBuilder.String())
-		}
-
-		// Track usage with error status
-		usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, 0)
-		s.trackUsageWithTokenUsage(c, usage, err)
-
-		// Send error event
-		errorChunk := map[string]interface{}{
-			"error": map[string]interface{}{
-				"message": err.Error(),
-				"type":    "stream_error",
-				"code":    "stream_failed",
-			},
-		}
-
-		errorJSON, marshalErr := json.Marshal(errorChunk)
-		if marshalErr == nil {
-			c.SSEvent("", string(errorJSON))
-		} else {
-			c.SSEvent("", map[string]interface{}{
-				"error": map[string]interface{}{
-					"message": "Failed to marshal error",
-					"type":    "internal_error",
-				},
-			})
-		}
-		flusher.Flush()
-		return
-	}
-
-	// Track successful streaming completion
-	// If no usage from stream, estimate it and send to client
-	if !hasUsage {
-		inputTokens, _ = token.EstimateInputTokens(req)
-		outputTokens = token.EstimateOutputTokens(contentBuilder.String())
-
-		// Use the first chunk ID, or generate one if not available
-		chunkID := firstChunkID
-		if chunkID == "" {
-			chunkID = fmt.Sprintf("chatcmpl-%d", time.Now().Unix())
-		}
-
-		// Send estimated usage as final chunk (only if not disabled)
-		if !disableStreamUsage {
-			usageChunk := map[string]interface{}{
-				"id":      chunkID,
-				"object":  "chat.completion.chunk",
-				"created": 0,
-				"model":   responseModel,
-				"choices": []map[string]interface{}{
-					{
-						"index":         0,
-						"delta":         map[string]interface{}{},
-						"finish_reason": nil,
-					},
-				},
-				"usage": map[string]interface{}{
-					"prompt_tokens":     inputTokens,
-					"completion_tokens": outputTokens,
-					"total_tokens":      inputTokens + outputTokens,
-				},
-			}
-
-			usageChunkJSON, err := json.Marshal(usageChunk)
-			if err == nil {
-				c.SSEvent("", usageChunkJSON)
-				flusher.Flush()
-			}
-		}
-	}
-
-	usage := protocol.NewTokenUsageWithCache(inputTokens, outputTokens, 0)
-	s.trackUsageWithTokenUsage(c, usage, nil)
-
-	// Send the final [DONE] message
-	// MENTION: must keep extra space
-	c.SSEvent("", " [DONE]")
-	flusher.Flush()
-}
-
-// convertChatCompletionToResponsesParams converts a chat completion request to responses API params
-func (s *Server) convertChatCompletionToResponsesParams(req *protocol.OpenAIChatCompletionRequest, actualModel string) responses.ResponseNewParams {
-	// Build input items from chat messages
-	inputItems := s.convertMessagesToResponseInputItems(req.Messages)
-
-	params := responses.ResponseNewParams{
-		Model:       actualModel,
-		Input:       responses.ResponseNewParamsInputUnion{OfInputItemList: responses.ResponseInputParam(inputItems)},
-		Temperature: req.Temperature,
-		TopP:        req.TopP,
-		MaxOutputTokens: func() param.Opt[int64] {
-			if req.MaxTokens.Valid() {
-				return param.NewOpt(req.MaxTokens.Value)
-			}
-			return param.Opt[int64]{}
-		}(),
-	}
-
-	// Add instructions from system message if present
-	instructionsFound := false
-	for _, msg := range req.Messages {
-		if !param.IsOmitted(msg.OfSystem) {
-			systemMsg := msg.OfSystem
-			if !param.IsOmitted(systemMsg.Content.OfString) {
-				params.Instructions = systemMsg.Content.OfString
-				instructionsFound = true
-				break
-			}
-		}
-	}
-
-	// If no system message (no instructions), add a default instruction
-	// This is required by ChatGPT backend API for Codex OAuth providers
-	if !instructionsFound {
-		params.Instructions = param.NewOpt("You are a helpful AI assistant.")
-	}
-
-	return params
-}
-
-// convertMessagesToResponseInputItems converts chat messages to response input items
-func (s *Server) convertMessagesToResponseInputItems(messages []openai.ChatCompletionMessageParamUnion) responses.ResponseInputParam {
-	var inputItems responses.ResponseInputParam
-
-	for _, msg := range messages {
-		switch {
-		case !param.IsOmitted(msg.OfUser):
-			userMsg := msg.OfUser
-			if !param.IsOmitted(userMsg.Content.OfString) {
-				content := userMsg.Content.OfString.Value
-				inputItem := responses.ResponseInputItemUnionParam{
-					OfMessage: &responses.EasyInputMessageParam{
-						Role: responses.EasyInputMessageRoleUser,
-						Content: responses.EasyInputMessageContentUnionParam{
-							OfString: param.NewOpt(content),
-						},
-					},
-				}
-				inputItems = append(inputItems, inputItem)
-			}
-
-		case !param.IsOmitted(msg.OfAssistant):
-			assistantMsg := msg.OfAssistant
-			if !param.IsOmitted(assistantMsg.Content.OfString) {
-				content := assistantMsg.Content.OfString.Value
-				inputItem := responses.ResponseInputItemUnionParam{
-					OfMessage: &responses.EasyInputMessageParam{
-						Role: responses.EasyInputMessageRoleAssistant,
-						Content: responses.EasyInputMessageContentUnionParam{
-							OfString: param.NewOpt(content),
-						},
-					},
-				}
-				inputItems = append(inputItems, inputItem)
-			}
-		}
-	}
-
-	return inputItems
-}
-
-func extractOpenAIMessages(messages []openai.ChatCompletionMessageParamUnion) []map[string]any {
-	if len(messages) == 0 {
-		return nil
-	}
-	b, _ := json.Marshal(messages)
-	var out []map[string]any
-	_ = json.Unmarshal(b, &out)
-	return out
 }
